requirements.txt
openai
python-dotenv
uuid
chromadb
streamlit
pypdf2
requests
beautifulsoup4
